requirement.txt
cryptography==2.1.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
numpy==1.15.4
pbr==5.1.1
protobuf==3.6.1
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pyxdg==0.25
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
SecretStorage==2.3.1
six==1.11.0
sklearn==0.0
subprocess32==3.5.3
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
